Struct postgres_range::RangeBound

pub struct RangeBound<S: BoundSided, T> {
    pub value: T,
    pub type_: BoundType,
    // some fields omitted
}

Represents a one-sided bound.

The side is determined by the S phantom parameter.

Fields

value

The value of the bound

type_

The type of the bound

Methods

impl<S, T> RangeBound<S, T> where S: BoundSided, T: PartialOrd

fn new(value: T, type_: BoundType) -> RangeBound<S, T>

Constructs a new range bound

fn in_bounds(&self, value: &T) -> bool

Determines if a value lies within the range specified by this bound.